Move charts folder into common
