+++ client/src/components/common/charts/CadenceTimeChart.js
@@ -1,43 +1,43 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ScatterChart from 'components/v2/ScatterChart';
-import { BLUE, GREY } from 'components/charts/colours';
+import { GREEN, GREY } from 'components/common/charts/colours';
 
 /**
  * @typedef {import('utils/timeSeries').TimeSeriesPoint} TimeSeriesPoint
  */
 
 /**
- * @typedef {object} PowerTimeChartProps
+ * @typedef {object} CadenceTimeChartProps
  * @property {TimeSeriesPoint[]} series Time series to render
  * @property {number} max The maximum value achieved
  */
 
 /**
- * Power-Time chart component
+ * Cadence-Time chart component
  *
- * @param {PowerTimeChartProps} props Props
- * @returns {React.Component<PowerTimeChartProps>} Component
+ * @param {CadenceTimeChartProps} props Props
+ * @returns {React.Component<CadenceTimeChartProps>} Component
  */
-export default function PowerTimeChart({ series, max }) {
+export default function CadenceTimeChart({ series, max }) {
   const data = series.map((point) => ({
     x: point.time / 1000,
     y: point.value,
   }));
   return (
     <ScatterChart
-      title="Power-Time"
+      title="Cadence-Time"
       xAxis={{ label: 'Time', unit: 's' }}
-      yAxis={{ label: 'Power', unit: 'W' }}
+      yAxis={{ label: 'Cadence', unit: 'RPM' }}
       data={data}
-      dataColour={BLUE}
+      dataColour={GREEN}
       max={max}
       maxColour={GREY}
     />
   );
 }
 
-PowerTimeChart.propTypes = {
+CadenceTimeChart.propTypes = {
   series: PropTypes.arrayOf(
     PropTypes.shape({
       time: PropTypes.number.isRequired,
