+++ client/src/components/v2/ScatterChart.stories.tsx
@@ -1,4 +1,4 @@
-import { BLUE, GREEN, GREY, PURPLE } from 'components/charts/colours';
+import { BLUE, GREEN, GREY, PURPLE } from 'components/common/charts/colours';
 import React from 'react';
 import { createStory, StoryTemplate } from 'utils/stories';
 import ScatterChart, { AxisProps, DataProps } from 'components/v2/ScatterChart';
